chore(apiaccess): Returning createdAt on keysearch
